Use 'show-body-only' Option for Fragments.
Tidylib already offers a built-in option ('show-body-only')
to return the contents of the body tag. Should use that rather
than a homebuilt regular expression (which is not the best way
to parse anything in html).

The level of indent is reduced in the returned html, so one
test needed to be updated as well.